+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## Release 1.0.0
 Initial release.
 - Allows a user to be designated as a "Proxy GM", used for executing certain actions normally restricted to GMs. The proxy GM must be logged on for said operations operations to execute correctly. The main GM can also function as the Proxy GM for campaigns where the game is not intended to be active outside of designated sessions.
+
+## Release 1.0.1
+- Copied the MasterGeeseLivingWorldTools dev environment/documentation to this repo.
+- Created Integration tests for the GMProxy setting.